Made Mosaic class the main export
